+++ source-hubspot-native/source_hubspot_native/api/line_items.py
@@ -33,8 +33,7 @@ def fetch_recent_line_items(
         page: PageCursor, count: int
     ) -> tuple[Iterable[TimestampedId], PageCursor]:
         return await fetch_search_objects(
-            Names.line_items, log, http, since, until, page,
-            should_crash_on_unordered_results=False,
+            Names.line_items, log, http, since, until, page
         )
 
     return fetch_changes_with_associations(