+++ source-hubspot-native/source_hubspot_native/api/orders.py
@@ -32,10 +32,7 @@ def fetch_recent_orders(
     async def do_fetch(
         page: PageCursor, count: int
     ) -> tuple[Iterable[TimestampedId], PageCursor]:
-        return await fetch_search_objects(
-            Names.orders, log, http, since, until, page,
-            should_crash_on_unordered_results=False,
-        )
+        return await fetch_search_objects(Names.orders, log, http, since, until, page)
 
     return fetch_changes_with_associations(
         Names.orders, Order, do_fetch, log, http, with_history, since, until